Create a new `BaseChatComponent` to handle shared functionalities and refactor `ChatComponent` to extend it
* **BaseChatComponent.tsx**
  - Create a new base component to handle shared functionalities
  - Move common state variables and methods from `ChatComponent.tsx`
  - Include methods for handling chat input, chat submission, and chat reset
  - Include methods for rendering chat threads and chat entry tab content

* **ChatComponent.tsx**
  - Extend `BaseChatComponent.tsx`
  - Remove duplicated state variables and methods
  - Use methods from `BaseChatComponent.tsx` for handling chat input, chat submission, and chat reset
  - Use methods from `BaseChatComponent.tsx` for rendering chat threads and chat entry tab content
